Introduce Utils\AccessAbstract
