+++ src/Resource/IdentifierCollection.php
@@ -2,12 +2,6 @@
 
 namespace Art4\JsonApiClient\Resource;
 
-use Art4\JsonApiClient\AccessInterface;
-use Art4\JsonApiClient\Utils\AccessTrait;
-use Art4\JsonApiClient\Utils\FactoryManagerInterface;
-use Art4\JsonApiClient\Exception\AccessException;
-use Art4\JsonApiClient\Exception\ValidationException;
-
 /**
  * A Identifier Collection holds only Identifier
  */
